[cast] Allow casting a checked pointer to a raw C pointer.
We did not allow casting a checked pointer to a raw C pointer before,
but actually the original Checked C allows casting a checked pointer
to a raw pointer; it is that the other way around is restricted in that
the compiler must be able to prove that the cast is safe.

Allowing this would simplify the implementation of extracting the
raw C pointer from a checked pointer: just a cast would do the work.
The current implementation has to go through memory to do the job.
See the current implementation at: https://github.com/jzhou76/checkedc-safe-mm-misc/blob/4239807e0df09436a1544e9300c8eb6fb61335f4/lib/safe_mm_checked.c#L283.