[cast] Fixed a bug in passing a cast raw pointer to a function call.
We now allow casting a checked pointer to a raw C pointer (https://github.com/jzhou76/checkedc-clang/commit/480e81761187bb85168c0ec00049ecf20168bfcb).
When passing a cast expression that casts a checked pointer to a raw C
pointer to a function call that expects to accept a raw pointer,
it should first extract the raw pointer from the checked pointer and
then do the cast and function call. See the code for more details.