[stack][WIP] Impl assigning an arith-expr of a local array to mmsafe ptr
We have decided to provide temporal memory safety for the stack.
The basic idea is to group all address-taken local objects into one
struct and put a lock at the beginning of the struct; and the compiler
creates mmsafe pointers to those objects with the key set to the lock
and the offset based on the offset of the pointee inside the struct.

On the clang side, an mmsafe pointer is created during IR generation
and its key is simply a placeholder, and the offset is the initial
offset calculated based on an arithmetic or address-of expression.
For example, for "p = buf + 3;", the initial offset for p is 3
(buf is a char array). The real key and final offset are set in an
LLVM pass after the initial IR generation.

Currently we have only implemented assigning an arithmetic expression
of a local array (including directly assigning the array, which is
equivalent to adding an index 0) to an mm_array_ptr.

The changes on the LLVM side: https://github.com/jzhou76/checkedc-llvm/commit/59c416ad3bc550dcb8324a2967f6eef085aee3fd

See the code for more details.