[stack] Fixed a bug in type checking involving addr-of expr of variables
We have a function that takes an address-of expression to check if it
should return an mmsafe pointer. The old implementation has a bug in
that it does not consider the LHS of the assignment (the addr-of
expression is the RHS of the assignment). If the RHS takes an address
of a local or global variable and the LHS is an mmsafe pointer, the
RHS should return an mmsafe pointer; but if the LHS is a raw pointer,
the RHS should return a raw pointer.

See the code for more details.