[global] Implemented assigning address-taken globals to mmsafe pointer.
Implemented the support for getting the address of a global variable
and assigning it to an mmsafe pointer. Before we required programmers
to add a "_checkable" qualifier to such variables, now this requirement
is no long needed. We use 1 as the key and create a lock for each
global pointed by mmsafe pointer(s). Grouping several globals into
one struct is possible but it is more difficult in terms of
implementation mainly because we need make sure all the globals in
a struct have the same attributes, e.g., whether they are constant.

One thing worthy of mention is that we implemented this support
in CodeGenModule. This is the correct place to go instead of putting
it in a Module pass. All module passes run after all function passes,
which means when a module pass is running, we may get code optimized
to a level that we do not want it to be. Actually all the Checkecd C
reparing code, i.e., CheckedCHarmonizeTypePass, CheckedCSecureStackPass,
and CheckedCAddLockToCheckablePass (should be fixed later), should
be moved to an earlier state when the initial IR code generation just
completes. Now we put the code of adding locks to address-taken globals
to the place that should be the earliest point after the initial IR
code generation.

TODO: Fix _checkable related code. Currently some of them conflicts
with the code for address-taken local/global variables.