Fixed a bug in IR code generation.
The code of handling address-taken global variables (https://github.com/jzhou76/checkedc-clang/commit/b7a1959e8facc4876cb4cefd56514034b53ab92a)
was added after the initial IR code generation. However, if there are
type or syntax errors happened before it, the compiling would not abort
immediately and the it would proceed to the address-taken globals
handling code, and it would crash the compiler because that code
executes based on the assumption that a complete IR module has been
built. The solution to this problem is to check if there are errors
before executing the addres-take global handling code. The changed
code in this commit is self-explanatory.