Impl. assigning an array in a struct pointed by a chcked ptr to a ptr.
Getting an array or an arithmetic expression of an array in a struct
pointed by a checked pointer should return an mmarray pointer. E.g.,
for "p->iarr + num;" where p is am mmptr to a struct and iarr is an
array in the struct, it should return an mmarrayptr that points to
p->iarr[num]. We did not realize this situation and simply allowed
assigning the result to a raw pointer. This commit fixed this problem.
It does type checking to disallow assining the result of such an
expression to a raw pointer, and during IR code generation, it generates
an mmarrayptr.

We found that em3d in the Olden benchmark suite has such code.
And lzfse also has several such cases.

All our benchmarks (Olden, thttpd, parson, and lzfse) can be compiled
by the current compiler and run successfully.