[variadic] Fixed the problem of passing an mmsafe ptr to a variadic fn.
We had this problem for quite a long time, and before our temporary
solution is to manually cast the mmsafe pointer argument to a raw ptr
before passing.

When passing an mmsafe ptr to a variadic function call, for some
(unknown) reason the generated type for the parameter is undefined,
while normally it would be "i8*" when passing a raw pointer.  We solve
this problem by simply ignoring the undefined type and set the argument
to be the value which is a raw ptr from an flatterned mmsafe ptr.

TODO: We should solve the "undefined type" problem earlier.