[WIP] Temporarily fixed a subtle bug in IR codegen.
The bug mentioned in the subject is described at: https://github.com/jzhou76/checkedc-llvm/commit/732b4ed10d3139a4f36a9ba90ad69ac3e622cb46.
Unfortunately that "fix" was incorrect (see https://github.com/jzhou76/checkedc-llvm/commit/ed4d32ba06b8411ff9e970ed6b883b571da9ac40).

I had a pass over all the usages of
Type::isMMPointerTy()/isMMArrayPointerTy()/isMMSafePointer(), and it
seems that they are not necessary. Some of them are redundant checks
if the type checks during semantic analysis are correct; e.g, during
pointer arithmetic codegen, we do not need to check if the pointer is
of MMArrayPtr type becasue only MMArrayPtr pointers are allowed to do
pointer arithmetic, and a compile error should be raised much earlier
if we perform an arithmetic on MMPtr. For some other cases, such as
during addres-of expression generation, we do not need the check methods
for pointer type from llvm::Type because the clang::Type for a target
pointer is available, which allows examining a pointer's type.  Further,
some other cases may not need these functions because we can rework the
code to achieve the same effect, as desmonstrated in this commit.

However, we may still need to check if a Value is of MMSafePtr type,
and we **mihgt** still need a way to distinguish the StructType for
an mmsafe pointer and the StructType that is defined by programmers but
happen to have exactly the same internal fields of mmsafe pointers.
If this is not needed, then simply checking if the Value is of
StructType which contains a pointer type and a i64 should suffice.

We should remove the isMMPtr/isMMArrayPtr of llvm::StructType
and refactor all code that replies on these two fields. This should be
feasible if we do not need to differentiate the StructType for
mmsafe pointers and a user-defined struct.