[global struct init] Impl. initializing a global struct with mmsafe ptr.
We did not implement initializing a global struct with mmsafe pointers
that are set to a string constant. The problem is that by default the
type of an mmsafe pointer field would be set to a raw pointer type.
This commit solved this problem by examining and fixing the type of a
global struct before a GloabalVariable is created, and also fixing the
initializer for the struct before the initializer is applied to the
struct. See the code for details.